stratos-backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
passlib[bcrypt]
python-jose
google-auth
sqlalchemy
psycopg2-binary
celery
redis
openai
astrapy
aiofiles
sse-starlette
reportlab
beautifulsoup4
httpx
